+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/barnetilsyn/beregning/TilsynBarnBeregningValideringUtil.kt
@@ -1,12 +1,11 @@
 package no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning
 
-import no.nav.tilleggsstonader.kontrakter.felles.overlapper
 import no.nav.tilleggsstonader.sak.felles.domain.BarnId
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.brukerfeilHvis
-import no.nav.tilleggsstonader.sak.infrastruktur.exception.feilHvis
 import no.nav.tilleggsstonader.sak.vedtak.TypeVedtak
 import no.nav.tilleggsstonader.sak.vedtak.UtgiftBeregningMåned
 import no.nav.tilleggsstonader.sak.vedtak.domain.VedtaksperiodeBeregning
+import no.nav.tilleggsstonader.sak.vedtak.validering.UtgifterValideringUtil.validerUtgifter
 
 object TilsynBarnBeregningValideringUtil {
     fun validerPerioderForInnvilgelse(
@@ -24,22 +23,6 @@ object TilsynBarnBeregningValideringUtil {
     private fun validerVedtaksperioder(vedtaksperioder: List<VedtaksperiodeBeregning>) {
         brukerfeilHvis(vedtaksperioder.isEmpty()) {
             "Kan ikke innvilge når det ikke finnes noen vedtaksperioder"
-        }
-    }
-
-    private fun validerUtgifter(utgifter: Map<BarnId, List<UtgiftBeregningMåned>>) {
-        brukerfeilHvis(utgifter.values.flatten().isEmpty()) {
-            "Kan ikke innvilge når det ikke finnes noen utgiftsperioder"
-        }
-        utgifter.entries.forEach { (_, utgifterForBarn) ->
-            feilHvis(utgifterForBarn.overlapper()) {
-                "Utgiftsperioder overlapper"
-            }
-
-            val ikkePositivUtgift = utgifterForBarn.firstOrNull { it.utgift < 0 }?.utgift
-            feilHvis(ikkePositivUtgift != null) {
-                "Utgiftsperioder inneholder ugyldig utgift: $ikkePositivUtgift"
-            }
         }
     }
 }
