+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/boutgifter/beregning/BoutgifterUtgiftService.kt
@@ -1,13 +1,10 @@
 package no.nav.tilleggsstonader.sak.vedtak.boutgifter.beregning
 
 import no.nav.tilleggsstonader.sak.felles.domain.BehandlingId
-import no.nav.tilleggsstonader.sak.infrastruktur.exception.feilHvis
-import no.nav.tilleggsstonader.sak.util.erFørsteDagIMåneden
-import no.nav.tilleggsstonader.sak.util.erSisteDagIMåneden
 import no.nav.tilleggsstonader.sak.vedtak.UtgiftBeregningDato
 import no.nav.tilleggsstonader.sak.vedtak.domain.TypeBoutgift
+import no.nav.tilleggsstonader.sak.vedtak.tilUtgiftBeregning
 import no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.VilkårService
-import no.nav.tilleggsstonader.sak.vilkår.stønadsvilkår.domain.Vilkår
 import org.springframework.stereotype.Service
 
 @Service
@@ -18,22 +15,5 @@ class BoutgifterUtgiftService(
         vilkårService
             .hentOppfylteBoutgiftVilkår(behandlingId)
             .groupBy { TypeBoutgift.fraVilkårType(it.type) }
-            .mapValues { (_, values) -> values.map { it.tilUtgiftBeregning() } }
-
-    private fun Vilkår.tilUtgiftBeregning(): UtgiftBeregningDato {
-        feilHvis(fom == null || tom == null || utgift == null) {
-            "Forventer at fra-dato, til-dato og utgift er satt. Gå tilbake til Vilkår-fanen, og legg til datoer og utgifter der. For utviklerteamet: dette gjelder vilkår=$id."
-        }
-        feilHvis(type.erLøpendeUtgifterBo() && !fom.erFørsteDagIMåneden()) {
-            "Noe er feil. Fom skal være satt til første dagen i måneden"
-        }
-        feilHvis(type.erLøpendeUtgifterBo() && !tom.erSisteDagIMåneden()) {
-            "Noe er feil. Tom skal være satt til siste dagen i måneden"
-        }
-        return UtgiftBeregningDato(
-            fom = fom,
-            tom = tom,
-            utgift = utgift,
-        )
-    }
+            .mapValues { (_, values) -> values.map { it.tilUtgiftBeregning() as UtgiftBeregningDato } }
 }
